refacto pages to components and a bit more
