--- interview/client/src/assets/css/Interview.css
@@ -1,334 +1,0 @@
-.interview-started {
-  /* display: grid;
-    grid-template-columns: 1fr 400px;
-    flex-wrap: wrap; */
-  display: flex;
-  min-height: 88.5vh;
-  height: auto;
-  flex-direction: column;
-  background-color: rgb(26, 25, 25);
-}
-
-/* TEXT */
-.chat-text {
-  box-sizing: border-box;
-  margin: 0 5px;
-  display: grid;
-  grid-template-rows: 1fr 80px;
-  border: 1px solid #444444;
-  border-radius: 5px;
-  padding: 0 5px;
-  visibility: visible;
-  width: 367px;
-  background-color: white;
-}
-
-.chat-text > .MuiInputBase-root.MuiOutlinedInput-root.MuiInputBase-formControl {
-  margin-bottom: 5px;
-}
-
-.title-chat-text {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-}
-
-.title-chat-text > span {
-  padding-left: 10px;
-  color: #757575;
-}
-
-.title-chat-text > #close-icon {
-  color: #757575;
-}
-
-.messagesList {
-  box-sizing: border-box;
-  padding: 10px;
-}
-
-.messageWriting {
-  /* flex: 1; */
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  align-items: center;
-  padding-bottom: 10px;
-}
-
-.messageForm {
-  display: flex;
-  align-items: center;
-}
-
-.message {
-  display: flex;
-  justify-content: space-between;
-}
-
-.messageDate {
-  font-style: italic;
-  justify-content: flex-end;
-}
-/* input chatText */
-.MuiInputBase-root {
-  width: 400px;
-}
-
-/* VIDEO */
-.chat-video {
-  height: 80vh;
-  display: grid;
-  /* grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: 1fr 30px; */
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr 10px;
-  width: 100%;
-}
-
-.chat-video > .icons {
-  display: none;
-}
-
-video {
-  width: 100% !important;
-  height: auto !important;
-  max-height: 80vh;
-  min-width: 100%;
-  min-height: 100%;
-  border-radius: 5px;
-}
-
-.partner-video {
-  overflow: hidden;
-  width: 80%;
-  height: 90%;
-  grid-column: span 2;
-  margin: auto;
-  border-radius: 5px;
-}
-.user-video {
-  overflow: hidden;
-  width: 200px;
-  z-index: 2;
-  grid-column: 2;
-  grid-row: 2;
-  height: 150px;
-  position: relative;
-  top: -140px;
-  padding: 0 10px 10px 0;
-}
-
-.MuiInputBase-root.MuiOutlinedInput-root.MuiInputBase-formControl {
-  width: 310px;
-}
-
-.container-video-chat {
-  flex-direction: row;
-  display: flex;
-  height: auto;
-}
-
-/* Users Room */
-
-.users-room {
-  box-sizing: border-box;
-  margin: 0 5px;
-  border: 1px solid #dfdfdf;
-  border-radius: 5px;
-  padding: 0 5px;
-  width: 367px;
-  /* display: none; */
-  background-color: white;
-}
-
-.title-users-room {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px;
-}
-
-.title-users-room > span {
-  padding-left: 10px;
-  color: #757575;
-}
-
-.title-users-room > #close-icon {
-  color: #757575;
-}
-
-/* Footer */
-.container-interview-footer {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  /*  padding-top: 10px; */
-  background: black;
-  align-items: center;
-}
-
-.container-interview-footer > .icons {
-  grid-column: 2/4;
-  margin: auto;
-  display: flex;
-}
-
-.container-interview-footer > .icons > div {
-  padding: 10px;
-}
-
-.footer-left-button {
-  display: flex;
-  justify-content: center;
-  color: #757575;
-  cursor: pointer;
-}
-
-.footer-left-button > .footer-icons {
-  padding: 10px;
-}
-
-svg {
-  margin: 0;
-}
-
-#chat-icon:focus,
-#group-icon:focus {
-  color: rgb(40 151 205);
-}
-
-#chat-icon:hover,
-#group-icon:hover {
-  color: rgb(204, 202, 202);
-  cursor: pointer;
-}
-
-#chat-icon,
-#group-icon {
-  background: none;
-  border: none;
-  color: rgb(160, 158, 158);
-}
-
-/* Modal */
-.box-modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 40%;
-  height: auto;
-  box-shadow: 2px 3px 13px 5px rgba(90, 90, 90, 0.47);
-  background: white;
-  border-radius: 5px;
-  padding: 10px;
-  box-sizing: border-box;
-}
-
-.header-modal {
-  display: flex;
-  justify-content: flex-end;
-  color: #a7a7a7;
-}
-
-.header-modal #close-icon {
-  color: #a7a7a7;
-}
-
-.content-modal {
-  text-align: center;
-}
-
-.text-modal {
-  color: #a7a7a7;
-}
-
-.footer-modal {
-  margin-top: 30px;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.footer-modal #cancel {
-  color: #a7a7a7;
-}
-
-.box-modal button {
-  margin: 0 10px;
-  font-weight: 500;
-}
-
-#modal-modal-title {
-  text-align: center;
-  color: #2c2c2c;
-}
-
-a {
-  text-decoration: none;
-}
-
-/* DISCONNECT PAGE */
-
-.container-disconnect {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.container-disconnect button {
-  margin: 0 5px;
-}
-
-.container-disconnect .buttons {
-  display: flex;
-  justify-content: center;
-}
-
-.container-disconnect h3 {
-  text-align: center;
-  color: #2c2c2c;
-}
-@media screen and (max-width: 1240px) {
-  .interview-started {
-    /* display: block; */
-    min-height: 90vh;
-  }
-  .container-video-chat {
-    flex-direction: column;
-  }
-  .chat-video {
-    height: 80vh;
-  }
-
-  .partner-video {
-    width: 100%;
-  }
-
-  .chat-video > .icons {
-    display: flex;
-    justify-content: center;
-    grid-column: span 3;
-  }
-
-  .chat-text {
-    /* margin-top: 30px;
-    height: 40vh; */
-  }
-
-  .container-interview-footer > .icons {
-    visibility: hidden;
-  }
-
-  .container-interview-footer {
-    display: none;
-  }
-
-  .users-room {
-    /* height: 40vh;
-    margin-top: 30px; */
-  }
-
-  .user-video {
-    top: -150px;
-  }
-}
