--- interview/client/src/assets/css/Message.css
@@ -1,4 +1,0 @@
-
-.messageText{
-    width: 50%;
-}
