+++ interview/client/src/components/LoggedOffPage/loggedOffPage.css
@@ -1,0 +1,20 @@
+.container-disconnect {
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.container-disconnect button {
+  margin: 0 5px;
+}
+
+.container-disconnect .buttons {
+  display: flex;
+  justify-content: center;
+}
+
+.container-disconnect h3 {
+  text-align: center;
+  color: #2c2c2c;
+}
