+++ interview/client/src/components/ModalLeaveInterview/modalLeaveInterview.css
@@ -1,0 +1,51 @@
+.box-modal {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 40%;
+  height: auto;
+  box-shadow: 2px 3px 13px 5px rgba(90, 90, 90, 0.47);
+  background: white;
+  border-radius: 5px;
+  padding: 10px;
+  box-sizing: border-box;
+}
+
+.header-modal {
+  display: flex;
+  justify-content: flex-end;
+  color: #a7a7a7;
+}
+
+.header-modal #close-icon {
+  color: #a7a7a7;
+}
+
+.content-modal {
+  text-align: center;
+}
+
+.text-modal {
+  color: #a7a7a7;
+}
+
+.footer-modal {
+  margin-top: 30px;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.footer-modal #cancel {
+  color: #a7a7a7;
+}
+
+.box-modal button {
+  margin: 0 10px;
+  font-weight: 500;
+}
+
+#modal-modal-title {
+  text-align: center;
+  color: #2c2c2c;
+}
