+++ interview/client/src/components/UsersSection/usersSection.css
@@ -1,0 +1,24 @@
+.users-room {
+  box-sizing: border-box;
+  margin: 0 5px;
+  border: 1px solid #dfdfdf;
+  border-radius: 5px;
+  padding: 0 5px;
+  width: 367px;
+  background-color: white;
+}
+
+.title-users-room {
+  display: flex;
+  justify-content: space-between;
+  padding: 10px;
+}
+
+.title-users-room > span {
+  padding-left: 10px;
+  color: #757575;
+}
+
+.title-users-room > #close-icon {
+  color: #757575;
+}
