+++ interview/client/src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './assets/css/index.css';
+import './index.css';
 import App from './App';
 // import { EndPointContext } from './useContext';
 import * as serviceWorker from './services/serviceWorker';
